// src/app/models/weather.model.ts
export interface Weather {
//  cod: number;
 //  message: string;
   // cnt: number;

   // list: [{}];
   //    dt: number;
   //  main: {
   //    temp: number,
   //    temp_min: number,
   //    temp_max: number,
   //    pressure: number,
   //    sea_level: number,
   //    grnd_level: number,
   //    humidity: number,
   //    temp_kf: number
   //  };
   //  weather: [{
   //    id: number,
   //    main: string,
   //    description: string,
   //    icon: number
   //  }];
   //  clouds: { all: number };
   //  wind: {
   //    speed: number,
   //    deg: number
   //  };
   //  sys: { pod: string };
    dt_txt: string;
    temp: number;
    wind: number;
    weather: string;
}